tests/wpt/web-platform-tests/orientation-event/t009-manual.html

<!DOCTYPE html>
<html>
  <head>
    <title>Rotate the device frame around its x axis</title>
    <meta name=viewport content="width=device-width, maximum-scale=1.0">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="author" title="Mosquito FP7">
    <link rel="reviewer author" title="Zhiqiang Zhang" href="mailto:zhiqiang.zhang@intel.com"> <!-- 2013-10-12 -->
  </head>
  <body>
    <p>Rotate the device frame around its x axis to run the test.</p>
    <div id="log"></div>
    <script>
      var beta = 180;
      var run = false;
      window.addEventListener("deviceorientation", function(e) {
        if (!run) {
          run = true;
          beta = e.beta;
        }
      }, false);

      test(function() {
        assert_true(beta > -180 && beta < 180);
      }, "Beta is between -180 and 180");

      test(function() {
        assert_false(beta > -2 && beta < 2);
      }, "Beta is not around zero (between -2 and 2)");
    </script>
  </body>
</html>
